file and env structure
