+++ app/index.js
@@ -1,3 +1,7 @@
+import * as dotenv from 'dotenv'
+
+dotenv.config()
+
 import addLocals               from './locals.js'
 import addRoutes               from './routes.js'
 import auth                    from '../middleware/auth.js'
@@ -6,15 +10,13 @@ import errors                  from '../middleware/errors.js'
 import express                 from 'express'
 import { fileURLToPath }       from 'url'
 import handleUncaughtException from './errors.js'
-import hbs                     from '../config/handlebars.js'
+import hbs                     from '../services/handlebars.js'
 import helmet                  from '../middleware/helmet.js'
 import locals                  from '../middleware/locals.js'
 import logger                  from '../middleware/logger.js'
 import path                    from 'path'
 import staticMiddleware        from '../middleware/static.js'
 import vary                    from '../middleware/vary.js'
-
-import { env, port } from '../config/app.js'
 
 const __filename = fileURLToPath(import.meta.url)
 const __dirname  = path.dirname(__filename)
@@ -27,7 +29,7 @@ const app = express()
 // Settings
 app.enable(`trust proxy`)
 app.engine(`hbs`, hbs.engine)
-app.set(`env`, env)
+app.set(`env`, process.env.NODE_ENV)
 app.set(`view engine`, `hbs`)
 app.set(`views`, path.join(__dirname, `../pages`))
 
@@ -46,6 +48,6 @@ app.use(logger)
 addRoutes(app.router)
 
 // Start server
-app.listen(port, () => {
-  console.info(`Server started on port ${ port } in ${ env } mode. Press Ctrl+C to terminate.`)
+app.listen(process.env.PORT, () => {
+  console.info(`Server started on port ${ process.env.PORT } in ${ process.env.NODE_ENV } mode. Press Ctrl+C to terminate.`)
 })
