--- config/database.js
@@ -1,5 +1,0 @@
-import Database from '../services/Database.js'
-
-const db = new Database
-
-export default db
