--- constants/index.js
@@ -1,3 +1,0 @@
-export const msAuthCookie = `AppServiceAuthSession`
-export const msAuthHeader = `X-MS-CLIENT-PRINCIPAL-NAME`
-export const msAuthUser   = `owner@digitallinguistics.io`
