+++ middleware/logger.js
@@ -1,8 +1,6 @@
-import { logging } from '../config/app.js'
-
 export default function logger(req, res, next) {
 
-  if (logging) {
+  if (process.env.LOGGING) {
     console.info(`${ new Date().toString() }: ${ req.originalUrl }`)
   }
