+++ package.json
@@ -54,6 +54,7 @@
     "citation-js": "^0.6.4",
     "clean-css": "^5.3.1",
     "cypress": "^12.1.0",
+    "dotenv": "^16.0.3",
     "esbuild": "^0.16.7",
     "esbuild-plugin-browserslist": "^0.6.0",
     "eslint": "^8.29.0",
