+++ pages/Error/Error-server.js
@@ -1,5 +1,3 @@
-import { env } from '../../config/app.js'
-
 export function PageNotFound(req, res) {
 
   if (req.method !== `GET`) {
@@ -13,7 +11,7 @@ export function PageNotFound(req, res) {
 
 export function ServerError(err, req, res, next) {
 
-  if (env !== `production`) console.error(err)
+  if (process.env.NODE_ENV !== `production`) console.error(err)
 
   res.error(`ServerError`)
