+++ pages/Languages/Languages-server.js
@@ -1,4 +1,4 @@
-import db from '../../config/database.js'
+import db from '../../services/database.js'
 
 import {
   hasAccess,
