+++ pages/Lexeme/Lexeme.test.js
@@ -1,5 +1,6 @@
-import { msAuthCookie } from '../../constants/index.js'
-import yamlParser       from 'js-yaml'
+import yamlParser from 'js-yaml'
+
+const msAuthCookie = Cypress.env(`msAuthCookie`)
 
 describe(`Lexeme page`, function() {
